Move DPLoot Masterkey Class to a separate file in the lib folder. This makes more sense as this class is used by all Triage classes in the project.
